Move ProfileContext out of windows/mod.rs.
